moved animationstuff from Graphics to Core
