+++ SluaghEngine/Core/Core.vcxproj
@@ -156,6 +156,7 @@
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="AnimationSystem.cpp" />
     <ClCompile Include="ITextManager.cpp" />
     <ClCompile Include="RenderableManagerInstancing.cpp" />
     <ClCompile Include="TextManager.cpp" />
@@ -209,6 +210,8 @@
     <ClInclude Include="..\includes\Core\IParticleSystemManager.h" />
     <ClInclude Include="..\includes\Core\IRenderableManager.h" />
     <ClInclude Include="..\includes\Core\ITransformManager.h" />
+    <ClInclude Include="AnimationStructs.h" />
+    <ClInclude Include="AnimationSystem.h" />
     <ClInclude Include="RenderableManagerInstancing.h" />
     <ClInclude Include="TextManager.h" />
     <ClInclude Include="AnimationManager.h" />
